requirements.txt
Django>=4.2
dnspython>=2.3
reportlab>=4.0
requests
